Removes the distance_field package
This package was copied from the old deprecated arm_navigation stack, and
copied to several new packages, including moveit.

Since some target in cmake remained, compilation from source is
incompatible.

Since in our application, we are not using the detector, this is package
is removed, and leave the one in moveit, since there it is likely to be
more important.